Remove JNLP references.
Mourn it as I may, disagree with Oracle's direction as I may, between Oracle
having stopped distribution of Java Web Start altogether, the certificate
"issues" of the years prior to that, the whole SecurityManager becoming
deprecated in JDK17, and the new launcher arguably being outright better, I
think it's safe to say at this point that JNLP isn't coming back. Probably at
all, and especially not to Haven.

It's a bit particularly melancholic for me since JNLP was the one original
reason to even write Haven in Java to begin with, but...

It's dead, Jim.