Re-enable boot-classpath warnings.
The warning is actually proper, and compiling without a boot-classpath from
Java 8 is likely to make the compiled product not work on Java 8 due to,
especially, automatically linking against chaining ByteBuffer methods that
did not exist in Java 8.

If anyone is reading this, the proper fix is to find the rt.jar file from
Java 8, and use build.properties to set the boot-classpath property to
point to it. Failing to do so will likely lead to the compiled product not
working on Java 8 (unless you are using an actual Java 8 JDK).