Allow the driver shader compiler to optimize away uniforms without crashing.
I have been hesitant to do this because optimized-away uniforms are usually
a sign of a problem rather than something to be silently ignored, but a
recent case with HueMod in a constant-white input convinced me otherwise.

I do think it is a bit weird that optimizing away uniform reads also
optimizes away the whole uniform, seeing as how it's part of the shader
interface, but apparently that's how it works.

Also not sure this is the nicest solution to the problem (see comment), but
it's unclear what the nicer solution would be.